pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "iconforge"
version = "0.1.0"
description = "IconForge backend service"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "fastapi>=0.111.0,<1.0.0",
    "uvicorn[standard]>=0.29.0,<0.30.0",
    "pydantic-settings>=2.2.1,<3.0.0",
    "aiofiles>=23.2.1,<24.0.0",
    "pillow>=10.3.0,<11.0.0",
    "numpy>=1.26.4,<2.0.0",
    "rembg>=2.0.57,<3.0.0",
    "python-multipart>=0.0.9,<0.0.10",
    "onnxruntime>=1.18.0,<1.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0,<9.0.0",
    "pytest-asyncio>=0.23.6,<0.24.0",
]

[tool.setuptools.packages.find]
where = ["app"]
include = ["app*"]
